+++ app/Domain/Stablecoin/Events/CollateralPositionUpdated.php
@@ -10,8 +10,8 @@ class CollateralPositionUpdated extends ShouldBeStored
 {
     public function __construct(
         public readonly string $position_uuid,
-        public readonly int $collateral_amount,
-        public readonly int $debt_amount,
+        public readonly float $collateral_amount,
+        public readonly float $debt_amount,
         public readonly float $collateral_ratio,
         public readonly string $status,
         public readonly array $metadata = []
